requirements.txt
opencv-python>=4.5.0
click>=8.0.0
tqdm>=4.65.0
openai-whisper>=20231117
moviepy>=1.0.3
Pillow>=9.0.0
gradio>=4.0.0
torch>=2.0.0
numpy>=1.21.0 